server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
bcrypt==3.2.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cloudinary==1.40.0
colorama==0.4.6
cryptography==46.0.5
dataclasses==0.6
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.23
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.4.0
orjson==3.11.7
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.7.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.4
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
starlette==0.37.2
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.30.1
watchfiles==1.1.1
websockets==16.0